% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievements_data.R
\name{achievements_data}
\alias{achievements_data}
\title{Provide dataframe of fake registry data}
\usage{
achievements_data(input, total_nasjonalt, lokal_enhet)
}
\value{
regData data frame
}
\description{
Provides a dataframe containing built-in data (and not a registry) for demo
purposes
}
