% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HealthUnitShortName.R
\name{get_HealthUnitShortName}
\alias{get_HealthUnitShortName}
\title{get HealthUnitShortName}
\usage{
get_HealthUnitShortName(UnitId, map_data)
}
\description{
Funksjon som gir navn på enhet basert på reaktivt ReshId (UnitId) basert
på innloggingsinfo i Falk. Denne funksjonen er basert på map_db_resh.
Denne kan brukes til plot f.eks., for å få rett enhet i tittel osv.
}
\details{
UnitId => user$org()
map_data => map_db_resh

Begge disse defineres i server.R
}
